+++ aten/src/ATen/ThreadLocalState.cpp
@@ -19,7 +19,7 @@ ThreadLocalState::ThreadLocalState()
 
   saved_tensors_default_hooks_ = at::SavedTensorDefaultHooks::get_stack();
 
-  torch_dispatch_mode_state_ = at::impl::TorchDispatchModeTLS::get_state();
+  torch_dispatch_mode_state_ = c10::impl::TorchDispatchModeTLS::get_state();
 }
 
 void ThreadLocalState::set_grad_mode(bool enabled) {
@@ -33,7 +33,7 @@ void ThreadLocalState::setThreadLocalState(
   // restore the dispatch key set TLS at the same time.
   c10::AutogradState::set_tls_state(state.autograd_tls_);
 
-  at::impl::TorchDispatchModeTLS::set_state(state.torch_dispatch_mode_state_);
+  c10::impl::TorchDispatchModeTLS::set_state(state.torch_dispatch_mode_state_);
 
   at::impl::PythonTorchFunctionTLS::set_state(state.python_torch_function_state_);
