+++ aten/src/ATen/ThreadLocalState.h
@@ -9,7 +9,7 @@
 
 #include <ATen/FuncTorchTLS.h>
 #include <ATen/PythonTorchFunctionTLS.h>
-#include <ATen/core/TorchDispatchModeTLS.h>
+#include <c10/core/impl/TorchDispatchModeTLS.h>
 #include <ATen/record_function.h>
 
 namespace at {
