+++ aten/src/ATen/core/PythonFallbackKernel.cpp
@@ -1,4 +1,4 @@
-#include <ATen/core/TorchDispatchModeTLS.h>
+#include <c10/core/impl/TorchDispatchModeTLS.h>
 #include <ATen/core/PythonFallbackKernel.h>
 #include <c10/core/SafePyObject.h>
 
@@ -51,7 +51,7 @@ void pythonFallback(const c10::OperatorHandle& op, torch::jit::Stack* stack) {
 
 
   // If Torch Dispatch Mode is active, use its PyInterpreter for dispatch
-  const auto& maybe_torch_dispatch_mode_state = at::impl::TorchDispatchModeTLS::get_state();
+  const auto& maybe_torch_dispatch_mode_state = c10::impl::TorchDispatchModeTLS::get_state();
   if (maybe_torch_dispatch_mode_state) {
     maybe_torch_dispatch_mode_state->pyinterpreter()->dispatch(op, stack);
     return;
