+++ aten/src/ATen/core/PythonFallbackKernel.h
@@ -1,5 +1,5 @@
 #pragma once
-
+#include <ATen/core/TorchDispatchUtils.h>
 
 namespace at {
 namespace impl {
