+++ torch/csrc/autograd/autograd_not_implemented_fallback.cpp
@@ -4,8 +4,9 @@
 
 #include <ATen/core/dispatch/Dispatcher.h>
 #include <ATen/core/ivalue.h>
+#include <ATen/core/TorchDispatchUtils.h>
 
-#include <ATen/core/TorchDispatchModeTLS.h>
+#include <c10/core/impl/TorchDispatchModeTLS.h>
 #include <torch/csrc/autograd/VariableTypeUtils.h>
 #include <torch/csrc/autograd/autograd.h>
 #include <torch/csrc/autograd/function.h>
